拆分订单修改：当订单中有供应链商品的情况下要拆分（即便只有供应链商品） 订单数据增加记录：发送到供应链系统的记录

+++ application/common/logic/CartLogic.php
@@ -912,7 +912,7 @@ class CartLogic extends Model
                 $cartList[$cartKey]['item_id'] = $specGoodsPrice['item_id'];
             } else {
                 $specGoodsPrice = SpecGoodsPrice::get(['goods_id' => $cart['goods_id']], '', false);
-                if ($specGoodsPrice) {
+                if ($specGoodsPrice && $specGoodsPrice['key'] != '') {
                     $cart->delete();
                     unset($cartList[$cartKey]);
                     continue;
